c-bindings: Fix the build
Fix the build of the old C bindings to the Haskell runtime. They will
probably not be used much now that there is a native C runtime, but it
was easy to get them to compile again:

build-gfctest.sh: Fix the path to QueryEng.gf to match the gf-contrib
                  repository layout. Also mark the script as executable.

PGFFFI.hs: Update the imports for the current Haskell standard.